Stores qc into separate file and uses NcML for aggregating
Instead of writing to the QC to the original data file directly, write
it to a separate file with an extension of .ncq.  Also write a simple
NcML file to define the aggregation between the data and the QC flags
and set up any relationships with `ancillary_variables` in the physical
variables which are being QCed.