Checks if files have already been QCed via QC file
Checks that files have already run QC for the configuration supplied
by checking the separate QC file.  Previously, QC was written to the
same file as the raw data, but with the redesign to move QC to a new
file, it was necessary to check the generated QC file instead.